Finish!!!

+++ src/Components/Card.jsx
@@ -8,25 +8,39 @@ const Card = ({ id, imageUrl, name, type }) => {
 
   const handleDelete = async (e) => {
     e.preventDefault();
-    try {
-        const response = await RestaurantService.deleteRestaurant(id); 
-        if (response.status === 200) {
-            Swal.fire({
-                title: "Restaurant Deleted",
-                text: response.data.message,
-                icon: "success", 
-            });
-            window.location.reload(); 
-        }
-    } catch (error) {
-        Swal.fire({
-            title: "Restaurant Deletion Failed",
-            text: error?.response?.data?.message || error.message,
-            icon: "error",
-        });
-    }
-};
 
+    const confirmDelete = await Swal.fire({
+      title: "Are you sure?",
+      text: "Do you really want to delete this restaurant? This action cannot be undone.",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+      cancelButtonText: "No, cancel!",
+    });
+
+    if (confirmDelete.isConfirmed) {
+      try {
+        const response = await RestaurantService.deleteRestaurant(id);
+        if (response.status === 200) {
+          Swal.fire({
+            title: "Restaurant Deleted",
+            text: response.data.message,
+            icon: "success",
+          }).then(() => {
+            window.location.reload();
+          });
+        }
+      } catch (error) {
+        Swal.fire({
+          title: "Restaurant Deletion Failed",
+          text: error?.response?.data?.message || error.message,
+          icon: "error",
+        });
+      }
+    }
+  };
 
   return (
     <div className="card card-compact w-72 bg-base-100 shadow-xl">
@@ -37,23 +51,24 @@ const Card = ({ id, imageUrl, name, type }) => {
         <h2 className="card-title">{name}</h2>
         <p>{type}</p>
 
-        {user && (user.roles.includes("ROLES_MODERATOR") || user.roles.includes("ROLES_ADMIN"))
-         && (
-        <div className="card-actions justify-end">
-          {user.roles.includes("ROLES_ADMIN") && (
-          <button
-            className="btn btn-error"
-            type="submit"
-            onClick={handleDelete}>
-            Delete
-          </button>
+        {user &&
+          (user.roles.includes("ROLES_MODERATOR") ||
+            user.roles.includes("ROLES_ADMIN")) && (
+            <div className="card-actions justify-end">
+              {user.roles.includes("ROLES_ADMIN") && (
+                <button
+                  className="btn btn-error"
+                  type="submit"
+                  onClick={handleDelete}
+                >
+                  Delete
+                </button>
+              )}
+              <a href={`/edit/${id}`} className="btn btn-primary">
+                Edit
+              </a>
+            </div>
           )}
-          <a href={`/edit/${id}`} className="btn btn-primary">
-            Edit
-          </a>
-        </div>
-        )}
-
       </div>
     </div>
   );
